+++ R/SQLite.R
@@ -1,3 +1,7 @@
+## usethis namespace: start
+#' @importFrom lifecycle deprecated
+#' @importFrom lifecycle deprecate_soft
+## usethis namespace: end
 #' @include SQLiteConnection.R
 #' @include SQLiteDriver.R
 NULL
@@ -70,3 +74,7 @@ check_vfs <- function(vfs) {
 #
 # This function checks for known protocols, or for a colon at the beginning.
 is_url_or_special_filename <- function(x) grepl("^(?:file|http|ftp|https|):", x)
+
+
+#' @import rlang
+NULL